Started creating help text
It will show syntax information by default, and allows the dev to
provide more detailed help text.

In the future it's worth considering an additional tool here that
will automatically take the description and information that the
dev has already provided for PowerCommandParser to drive some
documentation. This will be a very quick win for something to get
something written for a readme.

Personally I could use this quite a bit to turn any random project
into an official, nicely managed project. This will give the project
a chance to be used exterally and allow me to prioritize projects
based on which ones others are using etc.